FIX: POST w/ body instead of GET w/ querystr args
Right now arguments are encoded in the query string of the URL. This is
bad because when you send a large block of HTML this ends up being a
gigantic URL (and in my case, causes a HTTP 400 Bad Request). So instead
of sending a GET request with query string arguments, this fix now makes
it so that it sends a POST request with a body containing the JSON
encoded arguments.